+++ packages/panda-preset/src/slot-recipes/radio-card.ts
@@ -32,8 +32,7 @@ export const radioCardSlotRecipe = defineSlotRecipe({
         bg: "colorPalette.muted/20",
       },
       _disabled: {
-        opacity: "0.8",
-        borderColor: "border.disabled",
+        opacity: "0.5",
       },
       _checked: {
         zIndex: "1",
@@ -235,6 +234,11 @@ export const radioCardSlotRecipe = defineSlotRecipe({
             bg: "colorPalette.solid",
             color: "colorPalette.contrast",
             borderColor: "colorPalette.solid",
+          },
+        },
+        itemControl: {
+          _disabled: {
+            bg: "unset",
           },
         },
         itemIndicator: {