+++ packages/react/src/styled-system/token-transforms.ts
@@ -22,14 +22,15 @@ export const addConditionalCssVariables: TokenTransformer = {
   type: "value",
   name: "tokens/conditionals",
   transform(token, dictionary) {
-    const { prefix, formatCssVar } = dictionary
+    const { prefix, formatCssVar, getByName } = dictionary
 
     const refs = getReferences(token.value)
     if (!refs.length) return token.value
 
     refs.forEach((ref) => {
+      if (!getByName(ref)) return
       const variable = formatCssVar(ref.split("."), prefix)
-      token.value = token.value.replace(`{${variable.ref}}`, variable)
+      token.value = token.value.replaceAll(`{${ref}}`, variable.ref)
     })
 
     return token.value