+++ sandbox/panda-preset/panda.config.ts
@@ -9,5 +9,4 @@ export default defineConfig({
     extend: {},
   },
   outdir: "styled-system",
-  validation: "none",
 })